refactor: move `Network` to `p2`
