+++ crates/wasi/src/lib.rs
@@ -33,7 +33,7 @@ pub(crate) use self::ctx::WasiCtxBuilder;
 pub use self::error::{I32Exit, TrappableError};
 pub use self::filesystem::{DirPerms, FilePerms, OpenMode};
 pub use self::random::{Deterministic, thread_rng};
-pub use self::sockets::{Network, SocketAddrUse};
+pub use self::sockets::{AllowedNetworkUses, SocketAddrUse};
 #[doc(no_inline)]
 pub use async_trait::async_trait;
 #[doc(no_inline)]
