+++ crates/wasi/src/p2/mod.rs
@@ -244,7 +244,7 @@ mod write_stream;
 
 pub use self::ctx::{WasiCtx, WasiCtxBuilder};
 pub use self::filesystem::{FsError, FsResult};
-pub use self::network::{SocketError, SocketResult};
+pub use self::network::{Network, SocketError, SocketResult};
 pub use self::stdio::{
     AsyncStdinStream, AsyncStdoutStream, InputFile, IsATTY, OutputFile, Stderr, Stdin, StdinStream,
     Stdout, StdoutStream, stderr, stdin, stdout,
