+++ crates/wasi/src/p2/network.rs
@@ -1,5 +1,8 @@
-use crate::TrappableError;
+use core::net::SocketAddr;
+
 use crate::p2::bindings::sockets::network::{ErrorCode, Ipv4Address, Ipv6Address};
+use crate::sockets::SocketAddrCheck;
+use crate::{SocketAddrUse, TrappableError};
 
 pub type SocketResult<T> = Result<T, SocketError>;
 
@@ -41,4 +44,19 @@ pub(crate) fn to_ipv6_addr(addr: Ipv6Address) -> std::net::Ipv6Addr {
 pub(crate) fn from_ipv6_addr(addr: std::net::Ipv6Addr) -> Ipv6Address {
     let [x0, x1, x2, x3, x4, x5, x6, x7] = addr.segments();
     (x0, x1, x2, x3, x4, x5, x6, x7)
+}
+
+pub struct Network {
+    pub socket_addr_check: SocketAddrCheck,
+    pub allow_ip_name_lookup: bool,
+}
+
+impl Network {
+    pub async fn check_socket_addr(
+        &self,
+        addr: SocketAddr,
+        reason: SocketAddrUse,
+    ) -> std::io::Result<()> {
+        self.socket_addr_check.check(addr, reason).await
+    }
 }
